+++ datastore/tag/tag.go
@@ -10,6 +10,7 @@ package tag
 
 import (
 	"errors"
+	"github.com/MerinEREN/iiPackages/datastore/userTag"
 	"github.com/MerinEREN/iiPackages/session"
 	"golang.org/x/net/context"
 	"google.golang.org/appengine/datastore"
@@ -88,25 +89,27 @@ func PutAndGetMulti(s *session.Session, t *Tag) (Tags, error) {
 	return ts, err
 }
 
-// Delete removes the entity by the provided encoded entity key and returns an error.
+// Delete removes the entity and all the corresponding "userTag" entities in a transaction
+// by the provided encoded entity key
+// and returns an error.
 func Delete(ctx context.Context, ek string) error {
 	k, err := datastore.DecodeKey(ek)
 	if err != nil {
 		return err
 	}
-	return datastore.Delete(ctx, k)
-}
-
-// DeleteMulti removes the entitys by the provided encoded entity key slice
-// and returns an error.
-/* func DeleteMulti(s *session.Session, ekx []string) error {
-	var kx []*datastore.Key
-	for _, v := range ekx {
-		k, err := datastore.DecodeKey(v)
-		if err != nil {
-			return err
-		}
-		kx = append(kx, k)
+	kutx, err := userTag.GetKeysUserOrTag(ctx, k)
+	if err != nil {
+		return err
 	}
-	return datastore.DeleteMulti(s.Ctx, kx)
-}*/
+	opts := new(datastore.TransactionOptions)
+	opts.XG = true
+	err = datastore.RunInTransaction(ctx, func(ctx context.Context) (err1 error) {
+		err1 = datastore.DeleteMulti(ctx, kutx)
+		if err1 != nil {
+			return
+		}
+		err1 = datastore.Delete(ctx, k)
+		return
+	}, opts)
+	return err
+}
